docs: 파일명을 바꿀 땐 git mv를 씁시다
